upd to new unified map kit version

+++ library/src/com/idamob/map/details/MapItemDetailsController.java
@@ -23,7 +23,7 @@ public class MapItemDetailsController {
     private MapItemDetailsView detailsView;
 
     private IGeoPoint savedMapCenter;
-    private int savedMapZoom;
+    private float savedMapZoom;
     private boolean restoreLastMapPosition;
 
     private OnHideListener hideListener;
